+++ Sources/RAW/Macros.swift
@@ -5,6 +5,8 @@ import RAW_macros
 /// - arguments:
 /// 	- size_t: the byte count of the static buffer type. NOTE: this must be an integer literal in base10 format WITHOUT any special characters ('_' or others) in the value syntax
 @attached(member, 		names:			named(RAW_staticbuff_storetype),
+										named(RAW_access_staticbuff),
+										named(RAW_access),
 										named(init(RAW_staticbuff:)),
 										named(RAW_access_staticbuff_mutating),
 										named(RAW_encode(count:)),
@@ -17,6 +19,8 @@ public macro RAW_staticbuff(bytes:size_t) = #externalMacro(module:"RAW_macros", 
 
 @attached(member, 		names:			named(RAW_staticbuff_storetype),
 										named(init(RAW_staticbuff:)),
+										named(RAW_access_staticbuff),
+										named(RAW_access),
 										named(RAW_access_staticbuff_mutating),
 										named(RAW_encode(count:)),
 									 	named(RAW_encode(dest:)),
@@ -45,6 +49,7 @@ public macro RAW_staticbuff_binaryfloatingpoint_type<T:BinaryFloatingPoint>() = 
 
 @attached(member,		names:			named(init(_:)),
 										named(makeIterator()),
+										named(RAW_access),
 										named(RAW_access_mutating(_:)),
 										named(RAW_integer_encoding_impl),
 										named(RAW_convertible_unicode_encoding),